Rename nosto_tagging_product_index table to nosto_tagging_product_cache, remove redundant dependencies from indexers, rename IndexService to CacheService
